NumPy compatibility enhancements

+++ paddlespeech/s2t/models/wav2vec2/modules/wav2vec2_model.py
@@ -1476,7 +1476,7 @@ def compute_mask_indices(
                 lens = np.fromiter(
                     (e - s if e - s >= length + min_space else 0
                      for s, e in parts),
-                    np.int, )
+                    int, )
                 l_sum = np.sum(lens)
                 if l_sum == 0:
                     break
